LPS-119999 Move tests too.
